requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.4
astroid==2.5.6
attrs==21.2.0
backcall==0.2.0
black==21.5b2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
decorator==5.0.9
Django==3.2.4
djangorestframework==3.12.4
flake8==3.9.2
flake8-docstrings==1.6.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
packaging==20.9
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.9.0
pylint==2.8.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
snowballstemmer==2.1.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
